Update with changes on trunk to r13223
git-svn-id: file:///home/xapian-svn/xapian/branches/matchspy@13224 6b2f1b92-a4f3-0310-943f-de3cbe41748b

+++ xapian-bindings/python/pythontest3.py
@@ -23,7 +23,7 @@ import xapian
 import shutil
 import random
 
-from .testsuite import *
+from testsuite import *
 
 def setup_database():
     """Set up and return an inmemory database with 5 documents.
@@ -939,7 +939,7 @@ def test_postingsource():
             self.current = -1
 
         def get_termfreq_min(self): return 0
-        def get_termfreq_est(self): return self.max / 2
+        def get_termfreq_est(self): return int(self.max / 2)
         def get_termfreq_max(self): return self.max
         def next(self, minweight):
             self.current += 2
